+++ order-routing/rust/fulfillment-constraints/default/Cargo.toml.liquid
@@ -1,5 +1,5 @@
 [package]
-name = "{{name | replace: " ", "-" | downcase}}"
+name = "{{handle | replace: " ", "-" | downcase}}"
 version = "1.0.0"
 edition = "2021"
 rust-version = "1.62"
