+++ order-routing/rust/fulfillment-constraints/default/src/main.rs
@@ -1,24 +1,7 @@
-use shopify_function::prelude::*;
-use shopify_function::Result;
+use std::process;
+pub mod run;
 
-use serde::{Deserialize, Serialize};
-
-generate_types!(
-    query_path = "./input.graphql",
-    schema_path = "./schema.graphql"
-);
-
-#[derive(Serialize, Deserialize, Default, PartialEq)]
-struct Config {}
-
-#[shopify_function]
-fn function(_input: input::ResponseData) -> Result<output::FunctionResult> {
-    let operations = vec![];
-
-    // Build operations based on the input query response here.
-
-    Ok(output::FunctionResult { operations })
+fn main() {
+    eprintln!("Please invoke a named export.");
+    process::exit(1);
 }
-
-#[cfg(test)]
-mod tests;
